### 2021-10-22 99228fa
Fix(mf.py): flopy compatibility issues
