### 2021-10-22 147be8f
Update(ModflowAg): Set Accel term as optional variable
