+++ JOSS/paper.bib
@@ -59,7 +59,7 @@
          jounal = {Journal of Hydrology},
          volume = 537,
          pages = {45 - 60},
-         doi = {10.1016/j.hydrol.2016.03.026}
+         doi = {10.1016/j.jhydrol.2016.03.026}
 }
 
 @article{Gardner:2018,
