### 2021-10-25 a3044cf
Updated paper.bib and paper.md to include additional references
