### 2021-10-25 5eb8c5b
update Joss journal trigger
