### 2021-10-25 4205ea7
Added Jupyter DOI to paper.bib
