+++ JOSS/paper.bib
@@ -120,7 +120,7 @@
 	Organization = {IOS Press},
 	Pages = {87 - 90},
 	Title = {Jupyter Notebooks -- a publishing format for reproducible computational workflows},
-	doi = {10.3233/978-1-61499-649-1-87}
+	doi = {10.3233/978-1-61499-649-1-87},
 	Year = {2016}
 }
 
