### 2021-10-25 fdef8de
Update with changes from C. Stamos review comments
