+++ JOSS/paper.bib
@@ -220,7 +220,7 @@
       author={{Schroeder}, W. 
 	          and {Martin}, K. 
 			  and {Lorensen}, B.
-			  ~ISBN 978-1-930934-19-1},
+			  ~ISBN 978-1-930934-19-1.},
       url={https://books.google.com/books?id=rx4vPwAACAAJ},
       year={2006},
       publisher={Kitware}
