### 2021-10-26 #3852 @whedon
[REVIEW]: Integrated hydrologic model development and postprocessing for GSFLOW using pyGSFLOW
**Submitting author:** @jlarsen-usgs (<a href="http://orcid.org/0000-0002-1218-800X">Joshua Larsen</a>)
**Repository:** https://github.com/pygsflow/pygsflow.git
**Branch with paper.md** (empty if default branch):
**Version:** 1.0.2
