### 2021-10-29 a0f21a2
fix(PrmsDay.write()): fix directory not found error
