+++ .github/workflows/ci.yml
@@ -32,7 +32,7 @@ jobs:
       uses: actions/cache@v2.1.0
       env:
         # Increase this value to reset cache if environment.yml has changed
-        CACHE_NUMBER: 0
+        CACHE_NUMBER: 1
       with:
         path: ~/conda_pkgs_dir
         key: ${{ runner.os }}-${{ matrix.python-version }}-${{ matrix.run-type }}-${{ env.CACHE_NUMBER }}-${{ hashFiles('etc/environment.yml') }}
