+++ gsflow/modflow/mfag.py
@@ -168,9 +168,14 @@ class ModflowAg(flopy.modflow.ModflowAg):
                     OptionBlock.n_nested: 1,
                     OptionBlock.vars: OrderedDict(
                         [
-                            ("accel", {OptionBlock.dtype: float,
-                                       OptionBlock.nested: False,
-                                       OptionBlock.optional: True},)
+                            (
+                                "accel",
+                                {
+                                    OptionBlock.dtype: float,
+                                    OptionBlock.nested: False,
+                                    OptionBlock.optional: True,
+                                },
+                            )
                         ]
                     ),
                 },
@@ -1000,8 +1005,9 @@ class ModflowAg(flopy.modflow.ModflowAg):
                     nrec = len(t)
 
                     # check if this is block 16a
-                    if isinstance(options.tabfileswell, np.recarray) or \
-                            isinstance(options.tabfiles, np.recarray):
+                    if isinstance(
+                        options.tabfileswell, np.recarray
+                    ) or isinstance(options.tabfiles, np.recarray):
                         tf = True
                         well = ModflowAg.get_empty(nrec, block="tabfile_well")
                     else:
