### 2021-11-01 pygsflow/pygsflow#12 @thurber
JOSS review feedback
Hello `pygsflow` team! I'm a reviewer for your submission to JOSS tracked [here](https://github.com/openjournals/joss-reviews/issues/3852). I'll use this issue to track my feedback on the paper and software -- feel free to split out my comments into separate issues if that helps your workflow.

README.md

- [ ] #13

Paper

- [ ] Lines 48-49 in the paper have a duplicated expression "binds them into a single integrated model instance"
- [ ] Paragraph starting on line 62: since you mention external sensitivity analysis software, it could be useful for readers to cite the calibration software typically used in GSFLOW applications

Jupyter notebooks

- [x] #15
- [x] #16
- [x] #17
- [x] #14
- [x] Just wanted to say that the plots in the stream vector notebook are really nice looking!
