### 2021-11-03 #3852 @crvernon
:mega: Mid-week rally!  Looks like @thurber has started providing feedback to @jlarsen-usgs.  Let's keep things rolling towards a successful review!

:wave: @mdbartos - please update me to your progress thus far and don't hesitate to reach out if you have any questions.

👏 Keep up the good work!
