### 2021-11-04 95dbf4b
updates(notebook examples): update notebook examples
* reduce unneccessary notebook output pygsflow#14
* fix broken image links pygsflow#15
* remove comments and most teaching instructions from gsflow class pygsflow#16
* explain climate filtering process and add additional example pygsflow#17
