### 2021-11-04 1220c05
update miniconda cache number
* miniconda occasional fails to install on windows with caching, consider updating CI to exclude the cache action!
