### 2021-11-05 3502852
Update Readme.md
Generalize install release version (master) language

Closes #12
