### 2021-11-05 pygsflow/pygsflow#12 @jlarsen-usgs
Accidentally closed wrong issue.
