### 2021-11-05 bc33c63
update ci.yml to try to resolve miniconda/mamba cached installation issue on windows
