### 2021-11-10 #3852 @jlarsen-usgs
review from @thurber is linked to

pygsflow/pygsflow#12
