### 2021-11-10 pygsflow/pygsflow#12 @thurber
@jlarsen-usgs thanks for addressing these requests so quickly! A couple more small items:

- [ ] Could you elaborate a little bit more in the summary of pygsflow at the top of the README.md? This would be to satisfy the review requirement that there is a "Statement of Need" as part of the documentation (can be similar to what you have in the paper).
- [ ] Can you add a section to the README.md describing how to run the automated tests?
