### 2021-11-17 #3852 @crvernon
:mega: Mid-week rally!  Looks like issues are cycling with @thurber and @jlarsen-usgs.  Let's keep things rolling towards a successful review!

:wave: @mdbartos - Please let me know if you are unable to participate in this review at this time.  If you are still willing, could you reply here with an expected time of completion?  Thanks!
