### 2021-11-29 c16f497
Update paper.md and README.md
* removed duplicative text from paper.md #18
* Added section on running autotests to README.md

openjournals/joss-reviews#3852
