+++ README.md
@@ -47,6 +47,22 @@ Survey Software Release, 2 July 2021, https://doi.org/10.5066/P9NPZ5AD](https://
 ## Bugs
 The code is in active development and although there is a testing infrastructure set up we cannot catch all the software bugs without the help of users. If you find a bug or have an issue, please report it by opening a new issue. You can open a new issue by clicking the issues tab near the top of the page.
 
+## Running Autotests Locally
+pyGSFLOW uses github actions CI to automatically test code for each commit and pull request. These tests can also be run locally.
+To run tests locally from the pygsflow root directory, open a command prompt, anaconda prompt, or terminal window:
+
+with nosetests:
+```
+cd autotest
+nosetests -v
+```
+
+with pytest:
+```
+cd autotest
+pytest
+```
+
 ## Project History
 This project is a refinement and continuation of the original pygsflow repository at:
 
