### 2021-12-02 #3852 @thurber
hi @crvernon & @jlarsen-usgs, I've completed my review checklist and am satisfied with the resolutions of (pygsflow/pygsflow#12)[https://github.com/pygsflow/pygsflow/issues/12]. There's one unchecked box remaining, but i don't believe it to be critical path. Nice work!
