### 2021-12-02 #3852 @jlarsen-usgs
Thanks @thurber!

I have been waiting to update the text fully until all of the review comments are in (journal reviews and internal USGS). The unchecked box in pygsflow#12 will be addressed in the final version.
