### 2021-12-04 #3852 @mdbartos
I should be able to get through the checklist items this weekend, although there are some issues with the installation depending on which branch/python version I am using.

To be clear, should I be reviewing the `develop` branch of the current github repo?
