### 2021-12-04 pygsflow/pygsflow#19 @mdbartos
JOSS review feedback 2
Hi all, this checklist will contain feedback related to your [JOSS Review](https://github.com/openjournals/joss-reviews/issues/3852). I will continue to update this checklist as I move through my review:

## Installation

- [ ] Indicate which python versions are supported in the README.
- [ ] Installation fails from develop branch on Python3.6 with `EnvironmentError("pyGSFLOW is only supported with python 3.7 and above")`. Please indicate that Python3.6 is not supported in the README.
- [ ] Installation fails from develop branch on Python3.10 due to dependency error: `ERROR: Could not find a version that satisfies the requirement flopy>=3.3.5`
