### 2021-12-05 #3852 @mdbartos
Greetings, I've completed as much of my review as I can as of right now.

Currently I cannot install the software from the `develop` branch.

I can install from pypi and from the `master` branch. However, when I install this way, the unit tests are currently failing.

These issues need to be resolved before I can move further.

Thanks,
MDB
