### 2021-12-06 1e14a4e
Update(README.md): added requirements to installation instructions.
openjournals/joss-reveiws#3825
