### 2021-12-06 2eebcb8
update(README.md): added a Requirements section
#19
