+++ README.md
@@ -17,6 +17,10 @@ pyGSFLOW API documentation can be found @
 
 https://pygsflow.github.io/pygsflowdocs/
 
+## Examples
+Ipython notebook example problems can be found in the examples directory.  
+https://github.com/pygsflow/pygsflow/tree/master/examples
+
 ## Requirements
 Version 1.0.1 (Master branch and from pypi)
    1) Windows or Linux operating system (GSFLOW is not currently compiled for MacOS)  
@@ -27,10 +31,6 @@ Version 1.1.0 (Develop branch)
    1) Windows or Linux operating system (GSFLOW is not currently compiled for MacOS)  
    2) Python 3.7 or greater  
    3) FloPy 3.3.5 (`pip install -U https://github.com/modflowpy/flopy/zipball/develop`)
-
-## Examples
-Ipython notebook example problems can be found in the examples directory.  
-https://github.com/pygsflow/pygsflow/tree/master/examples
 
 ## Installation
 *Requirements*  
