### 2021-12-06 fa91d25
update(README.md): restructured README.md, updates to language
