### 2021-12-06 5b82905
update(README.md): more installation updates
