### 2021-12-06 #3852 @mdbartos
Welp, I was trying to install on a mac. So now it makes sense why the unit tests were failing on master.

Which branch does the JOSS release correspond to? Assuming it's `develop`, would it not make more sense to wait until an official release of flopy 3.3.5 is uploaded to pypi before publishing? Ideally, the installation procedure should be seamless for the end user. Especially if you are formally publishing a piece of software, you don't want users to be manually installing dependencies that are still in development, right?
