### 2021-12-06 786a5af
update(installation requirements): add support for flopy version 3.3.4
open-journals/joss-reviews#3852
