### 2021-12-06 0160261
update(ci, setup): add python 3.6 compatibility back into develop branch
