### 2021-12-06 775da1d
update ci to test py 3.6 with flopy 3.3.4
