### 2021-12-06 973890b
Add requirements for python==3.6
