### 2021-12-06 335ccb5
update version checks in setup.py
