### 2021-12-06 #3852 @jlarsen-usgs
@mdbartos

Instead of waiting until the next release of FloPy (which is most likely going to be sometime in late February), I went ahead and made some changes to `develop` that provide support for both FloPy 3.3.4 (current release) and FloPy's 3.3.5 develop version.
