### 2021-12-15 #3852 @crvernon
:mega: Mid-week rally!  Looks like we are getting close to completing the review!  @mdbartos and @jlarsen-usgs please post an update here on the status of what is remaining.  Thanks!
