### 2021-12-16 #3852 @mdbartos
Greetings,

I think I've finished the major part of my review. My only comments are the following:

- See the two minor comments on installation in my pygsflow repo issue.
- Include cleaned-up and simplified examples in the documentation pages (not just jupyter notebooks).

Aside from that, I think this is good to go.

Cheers,
MDB
