### 2021-12-16 #3852 @crvernon
✅ Thanks @mdbartos !

@jlarsen-usgs once you have addressed the reviewer comments let me know and I will confirm that all is in order and make suggestions for any edits that I think are necessary to satisfy journal requirements.
