### 2021-12-16 #3852 @jlarsen-usgs
Thanks for your review @mdbartos

I'll be working through the issues that you mentioned in the coming week or so!
