### 2021-12-29 0857a8c
update docstrings for pyGSFLOW documentation
pygsflow#20
