+++ gsflow/modflow/mf.py
@@ -20,10 +20,11 @@ class Modflow(fpModflow):
     namefile_ext : string, optional
         Extension for the namefile (the default is 'nam')
     version : string, optional
-        Version of MODFLOW to use (the default is 'mf2005').
+        Version of MODFLOW to use (the default is 'mfnwt')
+    version2 : str, optional
+        Secondary MODFLOW version to use (the default is 'gsflow')
     exe_name : string, optional
-        The name of the executable to use (the default is
-        'mf2005').
+        The name of the executable to use (the default is 'mfnwt').
     listunit : integer, optional
         Unit number for the list file (the default is 2).
     model_ws : string, optional
@@ -150,14 +151,26 @@ class Modflow(fpModflow):
 
     @property
     def version2(self):
+        """
+        Secondary version information about modflow
+
+        """
         return self._version2
 
     @property
     def model_ws(self):
+        """
+        Modflow model's base directory path
+
+        """
         return super(Modflow, self).model_ws
 
     @property
     def nper(self):
+        """
+        Returns the number of stress periods in the model
+
+        """
         try:
             return self.dis.nper
         except AttributeError:
@@ -165,6 +178,10 @@ class Modflow(fpModflow):
 
     @property
     def nrow(self):
+        """
+        Returns the number of model rows
+
+        """
         try:
             return self.modelgrid.nrow
         except AttributeError:
@@ -172,6 +189,10 @@ class Modflow(fpModflow):
 
     @property
     def ncol(self):
+        """
+        Returns the number of columns in the model
+
+        """
         try:
             return self.modelgrid.ncol
         except AttributeError:
@@ -179,6 +200,10 @@ class Modflow(fpModflow):
 
     @property
     def ncpl(self):
+        """
+        Returns the number of cells per layer in the model
+
+        """
         try:
             return self.modelgrid.ncpl
         except AttributeError:
@@ -197,9 +222,7 @@ class Modflow(fpModflow):
 
         Returns
         -------
-        val : list of strings
-            Can be used to see what packages are in the model, and can then
-            be used with get_package to pull out individual packages.
+        None
 
         """
         super(Modflow, self).change_model_ws(new_pth, reset_external)
@@ -297,10 +320,10 @@ class Modflow(fpModflow):
         f : str
             Path to MODFLOW name file to load.
         version : str, optional
-            MODFLOW version. Default 'mf2005', although can be modified on
+            MODFLOW version. Default 'mfnwt', although can be modified on
             loading packages unique to different MODFLOW versions.
         exe_name : str, optional
-            MODFLOW executable name. Default 'mf2005.exe'.
+            MODFLOW executable name. Default 'mfnwt.exe'.
         verbose : bool, optional
             Show messages that can be useful for debugging. Default False.
         model_ws : str
@@ -327,8 +350,8 @@ class Modflow(fpModflow):
         Examples
         --------
 
-        >>> import flopy
-        >>> ml = flopy.modflow.Modflow.load('model.nam')
+        >>> import gsflow
+        >>> ml = gsflow.modflow.Modflow.load('model.nam')
 
         """
 
@@ -612,3 +635,77 @@ class Modflow(fpModflow):
 
         # return model object
         return ml
+
+    def plot(self, SelPackList=None, **kwargs):
+        """
+        Plot 2-D, 3-D, transient 2-D, and stress period list (MfList)
+        model input data
+
+        Parameters
+        ----------
+        SelPackList : bool or list
+            List of of packages to plot. If SelPackList=None all packages
+            are plotted. (default is None)
+        **kwargs : dict
+            filename_base : str
+                Base file name that will be used to automatically generate file
+                names for output image files. Plots will be exported as image
+                files if file_name_base is not None. (default is None)
[…]
