### 2021-12-29 db62502
Update README.md and add additional linux installation instructions for using default GSFLOW binaries
pygsflow#19
Closes pygsflow#24
