### 2021-12-29 4435a5c
Update(README.md): updated requirements to add netcdf4
pygsflow#19
Closes pygsflow#25
