### 2021-12-30 1133501
Update(notebooks): continue long overdue notebook overhauls
