+++ gsflow/builder/__init__.py
@@ -6,3 +6,4 @@ from .builder_defaults import (
 )
 from .fishnet import GenerateFishnet
 from .modflow_builder import ModflowBuilder
+from .controlfile_builder import ControlFileBuilder
