+++ gsflow/builder/prms_builder.py
@@ -1,0 +1,191 @@
+import numpy as np
+from gsflow.builder import Defaults, PrmsDefaults
+from gsflow import PrmsParameters, ParameterRecord
+
+
+class PrmsBuilder(object):
+    """
+    Class for building PrmsParameter object using built in Defaults
+
+    PrmsBuilder builds a PrmsParameter object using default values
+    from gsflow.builder.Defaults or a user supplied Defaults object.
+
+    The user can then edit the PrmsParameter object to customize their
+    model runs and add climate, land cover, and soils information.
+
+    Parameters
+    ----------
+    stream_data_obj : gsflow.builder._StreamsObj
+        stream data from the flow accumulation stream creation method
+    crt_param_obj : gsflow.prms.PrmsParameters object
+        PrmsParameters object loaded from Crt() output
+    modelgrid : flopy.discretization.StructuredGrid object
+    hru_type : np.ndarray
+        hru_type array
+    defaults : gsflow.builder.Defaults object
+        optional gsflow.builder.Defaults object to build the PrmsParameters
+        object with.
+
+    """
+
+    def __init__(
+        self,
+        stream_data_obj,
+        crt_param_obj,
+        modelgrid,
+        hru_type=None,
+        defaults=None,
+    ):
+        if defaults is None:
+            self._defaults = Defaults().prms.to_dict()
+        elif isinstance(defaults, Defaults):
+            self._defaults = defaults.prms.to_dict()
+        elif isinstance(defaults, PrmsDefaults):
+            self._defaults = defaults.to_dict()
+        else:
+            raise TypeError("Defaults must be Default or PrmsDefault object")
+
+        self.stream_data_obj = stream_data_obj
+        self.crt_param_obj = crt_param_obj
+        self.modelgrid = modelgrid
+        self.hru_type = hru_type
+
+    def build(self, name=None):
+        """
+        Method to build a PrmsParameters object that can be used to build
+        a GSFLOW or PRMS model
+
+        Parameters
+        ----------
+        name : str
+            name of the parameter file
+
+        Returns
+        -------
+            gsflow.prms.parameters.PrmsParameters object
+
+        """
+        param_list = []
+
+        # parameter file defaults
+        paramfile_defaults = self._defaults["parameter"]
+
+        # load dimension defaluts
+        dimension_defaults = paramfile_defaults["dimensions"]
+
+        # get crt parameters??
+        crt_params = PrmsParameters(self.crt_param_obj)  # update this
+
+        # get the nhru from the model grid
+        nhru = self.modelgrid.nrow * self.modelgrid.ncol
+
+        # set segment
+        dimension_defaults[
+            "nsegment"
+        ] = (
+            self.stream_data_obj.iseg.max()
+        )  # I think this needs to be number of reaches!!!
+        dimension_defaults["ngw"] = nhru
+        dimension_defaults["nhru"] = nhru  # may change this
+        dimension_defaults["nssr"] = nhru
+        dimension_defaults[
+            "ncascade"
+        ] = nhru  # is this right??? I'm not so sure that this is. We should be able to get this from CRT...
+        dimension_defaults["ncascadegw"] = crt_params.get_values("ncascdgw")[0]
+        dimension_defaults["ndeplval"] = dimension_defaults["ndepl"] * 11
+        dimension_defaults["nsub"] = np.count_nonzero(np.unique(self.hru_type))
+        dimension_defaults["nhrucell"] = nhru
+        dimension_defaults["ndeplval"] = paramfile_defaults["dimensions"][
+            "ndeplval"
+        ]
+
+        for key, value in dimension_defaults.items():
+
+            param_record = ParameterRecord(
+                name=key, values=[value], datatype=1, file_name=name
[…]
