### 2022-01-04 11ffc79
update(autotests): update test_20.py
