+++ gsflow/builder/make_json_defaults.py
@@ -96,7 +96,7 @@ defaults = {
         "start_time": {"dtype": 1, "record": [1975, 10, 1, 0, 0, 0]},
         "rpt_days": {"dtype": 1, "record": 7},
         "cascade_flag": {"dtype": 1, "record": 1},
-        "cascadegw_flag": {"dtype": 1, "record": 1},
+        "cascadegw_flag": {"dtype": 1, "record": 2},
         "et_module": {"dtype": 4, "record": "potet_jh"},
         "precip_module": {"dtype": 4, "record": "precip_1sta"},
         "snarea_curve_flag": {"dtype": 1, "record": 0},
@@ -110,6 +110,8 @@ defaults = {
         "model_mode": {"dtype": 4, "record": "GSFLOW"},
         "csv_output_file": {"dtype": 4, "record": "gsflow.csv"},
         "gsflow_output_file": {"dtype": 4, "record": "gsflow.out"},
+        "executable_desc": {"dtype": 4, "record": "GSFLOW model"},
+        "executable_model": {"dtype": 4, "record": "prmsIV.exe"}
     },
     "parameter": {
         "dimensions": {
@@ -125,11 +127,12 @@ defaults = {
             "elev_units": {"dtype": 1, "dimension": "one", "record": 1},
             "precip_units": {"dtype": 1, "dimension": "one", "record": 1},
             "temp_units": {"dtype": 1, "dimension": "one", "record": 1},
+            "cascade_flg": {"dtype": 1, "dimension": "one", "record": 0},
             "cascade_tol": {"dtype": 2, "dimension": "one", "record": 5.0},
             "circle_switch": {"dtype": 1, "dimension": "one", "record": 1},
             "gw_pct_up": {
                 "dtype": 2,
-                "dimension": "ncascadegw",
+                "dimension": "ncascadgw",
                 "record": [1.0],
             },
             "hru_pct_up": {
@@ -139,26 +142,26 @@ defaults = {
             },
             "adjmix_rain": {
                 "dtype": 2,
-                "dimension": ["nhru", "nmonths"],
-                "record": [[1.0]] * 12,
+                "dimension": "nmonths",
+                "record": [1.0],
             },
             "ppt_zero_thresh": {"dtype": 2, "dimension": "one", "record": 0.0},
             "tmax_allsnow": {
                 "dtype": 2,
-                "dimension": ["nhru", "nmonths"],
-                "record": [[0.0]] * 12,
+                "dimension": "nmonths",
+                "record": [0.0],
             },
             "ppt_rad_adj": {
                 "dtype": 2,
-                "dimension": ["nhru", "nmonths"],
-                "record": [[0.02]] * 12,
+                "dimension": "nmonths",
+                "record": [0.02],
             },
             "radj_sppt": {"dtype": 2, "dimension": "nhru", "record": [0.44]},
             "radj_wppt": {"dtype": 2, "dimension": "nhru", "record": [0.5]},
             "radmax": {
                 "dtype": 2,
-                "dimension": ["nhru", "nmonths"],
-                "record": [[0.8]] * 12,
+                "dimension": "nmonths",
+                "record": [0.8],
             },
             "potet_sublim": {"dtype": 2, "dimension": "nhru", "record": [0.5]},
             "rad_trncf": {"dtype": 2, "dimension": "nhru", "record": [0.5]},
@@ -175,8 +178,8 @@ defaults = {
             "albset_snm": {"dtype": 2, "dimension": "one", "record": 0.2},
             "cecn_coef": {
                 "dtype": 2,
-                "dimension": ["nhru", "nmonths"],
-                "record": [[5.0]] * 12,
+                "dimension": "nmonths",
[…]
