+++ gsflow/builder/prms_builder.py
@@ -16,9 +16,9 @@ class PrmsBuilder(object):
     Parameters
     ----------
     stream_data_obj : gsflow.builder._StreamsObj
-        stream data from the flow accumulation stream creation method
-    crt_param_obj : gsflow.prms.PrmsParameters object
-        PrmsParameters object loaded from Crt() output
+        stream data from the flow accumulation make_streams() method
+    cascade_obj : gsflow.builder._Cascades object
+        cascade data object from the flow accumulation get_cascades() method
     modelgrid : flopy.discretization.StructuredGrid object
     hru_type : np.ndarray
         hru_type array
@@ -31,9 +31,11 @@ class PrmsBuilder(object):
     def __init__(
         self,
         stream_data_obj,
-        crt_param_obj,
+        cascades_obj,
         modelgrid,
+        dem,
         hru_type=None,
+        hru_subbasin=None,
         defaults=None,
     ):
         if defaults is None:
@@ -46,11 +48,13 @@ class PrmsBuilder(object):
             raise TypeError("Defaults must be Default or PrmsDefault object")
 
         self.stream_data_obj = stream_data_obj
-        self.crt_param_obj = crt_param_obj
+        self.cascades_obj = cascades_obj
         self.modelgrid = modelgrid
+        self.dem = dem
         self.hru_type = hru_type
+        self.hru_subbasin = hru_subbasin
 
-    def build(self, name=None):
+    def build(self, name=None, area_conv=2.47105e-4):
         """
         Method to build a PrmsParameters object that can be used to build
         a GSFLOW or PRMS model
@@ -59,6 +63,8 @@ class PrmsBuilder(object):
         ----------
         name : str
             name of the parameter file
+        area_conv : float
+            conversion factor from model units sq. to arces
 
         Returns
         -------
@@ -73,31 +79,30 @@ class PrmsBuilder(object):
         # load dimension defaluts
         dimension_defaults = paramfile_defaults["dimensions"]
 
-        # get crt parameters??
-        crt_params = PrmsParameters(self.crt_param_obj)  # update this
-
         # get the nhru from the model grid
         nhru = self.modelgrid.nrow * self.modelgrid.ncol
 
-        # set segment
-        dimension_defaults[
-            "nsegment"
-        ] = (
-            self.stream_data_obj.iseg.max()
-        )  # I think this needs to be number of reaches!!!
+        # set segment and reach
+        dimension_defaults["nsegment"] = self.stream_data_obj.iseg.max()
+        dimension_defaults["nreach"] = self.stream_data_obj.reach_data.size
+
         dimension_defaults["ngw"] = nhru
-        dimension_defaults["nhru"] = nhru  # may change this
+        dimension_defaults["ngwcell"] = nhru
+        dimension_defaults["nhru"] = nhru
+        dimension_defaults["nhrucell"] = nhru
         dimension_defaults["nssr"] = nhru
-        dimension_defaults[
-            "ncascade"
-        ] = nhru  # is this right??? I'm not so sure that this is. We should be able to get this from CRT...
-        dimension_defaults["ncascadegw"] = crt_params.get_values("ncascdgw")[0]
+        dimension_defaults["ncascade"] = self.cascades_obj.ncascade
+        dimension_defaults["ncascadgw"] = self.cascades_obj.ncascade
         dimension_defaults["ndeplval"] = dimension_defaults["ndepl"] * 11
         dimension_defaults["nsub"] = np.count_nonzero(np.unique(self.hru_type))
         dimension_defaults["nhrucell"] = nhru
         dimension_defaults["ndeplval"] = paramfile_defaults["dimensions"][
             "ndeplval"
         ]
+        try:
+            dimension_defaults["ngwcell"] = self.modelgrid.nnodes
+        except TypeError:
[…]
