### 2022-01-06 #3852 @crvernon
:mega: Mid-week rally!

It would be good to get this one wrapped up soon @jlarsen-usgs so we can free up our reviewers.  How are things looking on your side of things for finishing up the remaining issues?  It looks like @mdbartos has two boxes left to check off and @thurber has approved all.

Thanks!
