### 2022-01-06 #3852 @jlarsen-usgs
Hi @crvernon

I've advanced it to the last step in the USGS Bureau Approval process that goes in conjunction with the journal review. Once I get the approval there I'll be able to check the remaining issues and have the work on my end wrapped up. Thanks for checking in, the final step on my end shouldn't take very long (< 1 week).
