### 2022-01-06 #3852 @crvernon
Sounds great @jlarsen-usgs !  Let me know when you are clear and I'll make my final pass and recommendation.
