+++ gsflow/param_base.py
@@ -69,7 +69,7 @@ class ParameterBase(object):
             return record
         else:
             err = "The record does not exist..."
-            warnings.warn(err, UserWarning)
+            warnings.warn_explicit(err, UserWarning, "param_base.py", 72)
             return None
 
     def get_values(self, name):
@@ -90,7 +90,7 @@ class ParameterBase(object):
 
         if record is None:
             err = "The record does not exist..."
-            warnings.warn(err, UserWarning)
+            warnings.warn_explicit(err, UserWarning, "param_base.py", 93)
             return None
 
         elif isinstance(record.values, np.ndarray):
@@ -98,7 +98,7 @@ class ParameterBase(object):
 
         else:
             err = "The values does not exist..."
-            warnings.warn(err, UserWarning)
+            warnings.warn_explicit(err, UserWarning, "param_base.py", 101)
             return None
 
     def set_values(self, name, values):
@@ -118,7 +118,7 @@ class ParameterBase(object):
 
         if record is None:
             err = "The record does not exist {}".format(name)
-            warnings.warn(err, UserWarning)
+            warnings.warn_explicit(err, UserWarning, "param_base.py", 121)
             return None
 
         elif isinstance(record.values, np.ndarray):
@@ -127,7 +127,7 @@ class ParameterBase(object):
 
         else:
             err = "The record does not exist..."
-            warnings.warn(err, UserWarning)
+            warnings.warn_explicit(err, UserWarning, "param_base.py", 130)
             return None
 
     def _check_before_add(self, name, values, replace=False):
@@ -160,7 +160,7 @@ class ParameterBase(object):
                 )
             )
             if not replace:
-                warnings.warn(err, UserWarning)
+                warnings.warn_explicit(err, UserWarning, "param_base.py", 163)
             return False
 
         return True
@@ -208,9 +208,8 @@ class ParameterBase(object):
             raise ValueError("Record name must be a string")
 
         if name not in self.record_names:
-            warnings.warn(
-                "The record does not exist: {}".format(name), UserWarning
-            )
+            err = f"The record does not exist: {name}"
+            warnings.warn_explicit(err, UserWarning, "param_base.py", 212)
             return
 
         for index, nm in enumerate(self.record_names):
