### 2022-01-11 1be292a
updates(path print statements) remove absolute path print statements
