### 2022-01-11 82dc626
update(modsim.py): update warning message
