+++ JOSS/paper.bib
@@ -1,0 +1,197 @@
+@article{Bakker:2016,
+        author = {{Bakker}, M.
+                  and {Post}, V.
+                  and {Langevin}, C.~D.
+                  and {Hughes}, J.~D.
+                  and {White}, J.~T.
+                  and {Starn}, J.~J.
+                  and {Fienen}, M.~N},
+        title = {Scripting MODFLOW Model Development Using Python and Flopy},
+        journal = {Groundwater},
+        volume = 54,
+        pages = {733 - 739},
+        year = 2016,
+        doi = {10.1111/gwat.12413}
+}
+
+@misc{Bakker:2021,
+      author = {{Bakker}, M.
+                and {Post}, V.
+                and {Langevin}, C.~D.
+                and {Hughes}, J.~D.
+                and {White}, J.~T.
+                and {Leaf}, A.~T.
+                and {Paulinski}, S.~R.
+                and {Larsen}, J.~D.
+                and {Towes}, M.
+                and {Morway}, E.~D.
+                and {Bellino}, J.~C.
+                and {Starn}, J.~J.
+                and {Fienen}, M.~N.},
+      title = {FloPy v3.3.3},
+      year = {2021},
+      publisher = {U.S. Geological Survey Software Release},
+      doi = {10.5066/F7BK19FH}
+}
+
+@article{Fatichi:2016,
+         author = {{Fatichi}, S.
+                   and {Vivoni}, E.~R.
+                   and {Ogden}, F.~L.
+                   and {Ivanov}, V.~Y.
+                   and {Mirus}, B.
+                   and {Gochis}, D.
+                   and {Downer}, C.~W.
+                   and {Camporese}, M.
+                   and {Davison}, J.~H.
+                   and {Ebel}, B.
+                   and {Jones}, N.
+                   and {Kim}, J.
+                   and {Mascaro}, G.
+                   and {Niswonger}, R.
+                   and {Restrepo}, P.
+                   and {Rigon}, R.
+                   and {Shen}, C.
+                   and {Sulis}, M.
+                   and {Tarboton}, D.},
+         title = {An overview of current applications, challenges, and future trends in distributed process-based models in hydrology},
+         year = {2016},
+         jounal = {Journal of Hydrology},
+         volume = 537,
+         pages = {45 - 60},
+         doi = {10.1016/j.hydrol.2016.03.026}
+}
+
+@article{Gardner:2018,
+         author = {{Gardner}, M.~A.
+                   and {Morton}, C.~G.
+                   and {Huntington}, J.~L.
+                   and {Niswonger}, R.~G.
+                   and {Henson}, W.~R.},
+         title = {Input data processing tools for the integrated hydrologic model GSFLOW.},
+         journal = {Environmental modelling & software},
+         year = {2018},
+         volume = 109,
+         pages = {41 - 53},
+         doi = {10.1016/j.envsoft.2018.07.020}
+}
+
+@book{Harbaugh:2005,
+      author = {{Harbaugh}, A.~W.},
+      Booktitle = {U.S. Geological Survey Techniques and Methods, book 6},
+      Publisher = {U.S. Geological Survey},
+      Title = {MODFLOW-2005, the U.S. Geological Survey modular ground-water model -- the Ground-Water Flow Process},
+      year = 2005,
+      doi = {10.3133/tm6A16}
+}
+
+@misc{Labadie:2006,
+      author = {{Labadie}, J.
+                and {Larson}, R.},
+      title = {MODSIM: River basin management decision support system, User Manual},
+      year = {2006},
[…]
