+++ gsflow/output/discretization.py
@@ -1,0 +1,297 @@
+import numpy as np
+import shapefile
+from matplotlib.patches import Polygon
+from matplotlib.collections import PatchCollection
+import matplotlib.pyplot as plt
+
+
+class PrmsDiscretization(object):
+    """
+    The PrmsDiscretization object is for plotting hru's and
+    hru data from Prms on matplotlib plots.
+
+    Could also be used for exporting shape files of data
+
+    Parameters
+    ----------
+    xypts : list
+        list of hru [(x1, y1)....(xn, yn)] points
+        3 dimensional
+
+    Examples
+    --------
+
+    load from shapefile
+
+    >>> dis = gsflow.output.PrmsDiscretization.load_from_shapefile("myshape.shp")
+
+    load from flopy grid
+
+    >>> gsf = gsflow.GsflowModel.load_from_file("mycontrol.control")
+    >>> ml = gsflow.mf
+    >>> dis = gsflow.output.PrmsDiscretization.load_from_flopy(model=ml)
+
+    """
+
+    def __init__(self, xypts):
+        self._xypts = xypts
+        self._nhru = len(xypts)
+        self._xcenters = None
+        self._ycenters = None
+
+        xmin, xmax, ymin, ymax = (None, None, None, None)
+        for hru in xypts:
+            hru = np.array(hru)
+            if xmin is None:
+                xmin = np.min(hru.T[0])
+                xmax = np.max(hru.T[0])
+                ymin = np.min(hru.T[1])
+                ymax = np.max(hru.T[1])
+
+            else:
+                if np.min(hru.T[0]) < xmin:
+                    xmin = np.min(hru.T[0])
+                if np.max(hru.T[0]) > xmax:
+                    xmax = np.max(hru.T[0])
+                if np.min(hru.T[1]) < ymin:
+                    ymin = np.min(hru.T[1])
+                if np.max(hru.T[1]) > ymax:
+                    ymax = np.max(hru.T[1])
+
+        self._extent = (xmin, xmax, ymin, ymax)
+
+    @property
+    def x_hru_centers(self):
+        """
+        Returns
+        -------
+            np.ndarray of x-centers for each hru
+
+        """
+        if self._xcenters is None:
+            self._get_centers()
+
+        return self._xcenters
+
+    @property
+    def y_hru_centers(self):
+        """
+        Returns
+        -------
+            np.ndarray of y-centers for each hru
+
+        """
+        if self._ycenters is None:
+            self._get_centers()
+
+        return self._ycenters
+
+    @property
+    def nhru(self):
+        """
+        Returns
+        -------
+            number of hrus
+
+        """
+        return self._nhru
+
+    @property
+    def xypts(self):
+        """
+        Returns
+        -------
+            np.ndarray of xy-points for each hru
+
+        """
+        return self._xypts
+
+    @property
+    def extent(self):
+        """
+        Returns
+        -------
+            tuple (xmin, xmax, ymin, ymax)
+        """
+        return self._extent
+
+    def _get_centers(self):
+        """
+        Method to get the mean center of the grid cell.
+        This method is limited and will eventually need to
+        be updated for complex shapes.
+
+        """
+        xc = []
+        yc = []
+        for hru in self._xypts:
+            hru = np.array(hru)
+            xc.append(np.mean(hru.T[0]))
+            yc.append(np.mean(hru.T[1]))
+
+        self._xcenters = xc
+        self._ycenters = yc
+
+    def get_hru_points(self, hru):
+        """
+        Get the x, y coordinate points for a hru
+
+        Parameters
+        ----------
+        hru : int
+            hru number
+
+        Returns
+        -------
+            list of x, y coordinate points
+
+        """
+        return self._xypts[hru - 1]
+
+    @staticmethod
[…]
