+++ gsflow/modflow/mf.py
@@ -1,0 +1,585 @@
+import os
+import inspect
+import flopy
+import platform
+from flopy.utils import TemporalReference
+from flopy.utils.reference import SpatialReference
+from flopy.modflow.mf import Modflow as fpModflow
+from ..utils import mfreadnam
+
+
+class Modflow(fpModflow):
+    """
+    GSFLOW overridden MODFLOW Model Class. This class is a child of
+    the FloPy flopy.modflow.Modflow class
+
+    Parameters
+    ----------
+    modelname : string, optional
+        Name of model.  This string will be used to name the MODFLOW input
+        that are created with write_model. (the default is 'modflowtest')
+    namefile_ext : string, optional
+        Extension for the namefile (the default is 'nam')
+    version : string, optional
+        Version of MODFLOW to use (the default is 'mf2005').
+    exe_name : string, optional
+        The name of the executable to use (the default is
+        'mf2005').
+    listunit : integer, optional
+        Unit number for the list file (the default is 2).
+    model_ws : string, optional
+        model workspace.  Directory name to create model data sets.
+        (default is the present working directory).
+    external_path : string
+        Location for external files (default is None).
+    verbose : boolean, optional
+        Print additional information to the screen (default is False).
+    load : boolean, optional
+         (default is True).
+    silent : integer
+        (default is 0)
+
+    Attributes
+    ----------
+
+    Methods
+    -------
+
+    See Also
+    --------
+
+    Notes
+    -----
+
+    Examples
+    --------
+
+    >>> import flopy
+    >>> m = flopy.modflow.Modflow()
+
+    """
+
+    def __init__(
+        self,
+        modelname="modflowtest",
+        namefile_ext="nam",
+        version="mfnwt",
+        version2="gsflow",
+        exe_name="mfnwt.exe",
+        structured=True,
+        listunit=2,
+        model_ws=".",
+        external_path=None,
+        verbose=False,
+        **kwargs
+    ):
+
+        if platform.system().lower() == "windows":
+            if not exe_name.endswith(".exe"):
+                exe_name += ".exe"
+
+        else:
+            if exe_name.endswith(".exe"):
+                exe_name = exe_name[:-4]
+
+        super(Modflow, self).__init__(
+            modelname=modelname,
+            namefile_ext=namefile_ext,
+            version=version,
+            exe_name=exe_name,
+            structured=structured,
+            listunit=listunit,
+            model_ws=model_ws,
+            external_path=external_path,
+            verbose=verbose,
+            **kwargs
+        )
+
+        self.version2 = version2
+
+        # Using a local gsflow import to avoid circular imports with python 2.7
+        import gsflow
+
+        # we can override packages here by setting the package object
+        # a custom written package if flopy doesn't work with pygsflow
+        self.mfnam_packages = {
+            "zone": flopy.modflow.ModflowZon,
+            "mult": flopy.modflow.ModflowMlt,
+            "pval": flopy.modflow.ModflowPval,
+            "bas6": flopy.modflow.ModflowBas,
+            "dis": flopy.modflow.ModflowDis,
+            "disu": flopy.modflow.ModflowDisU,
+            "bcf6": flopy.modflow.ModflowBcf,
+            "lpf": flopy.modflow.ModflowLpf,
+            "hfb6": flopy.modflow.ModflowHfb,
+            "chd": flopy.modflow.ModflowChd,
+            "fhb": flopy.modflow.ModflowFhb,
+            "wel": flopy.modflow.ModflowWel,
+            "mnw1": flopy.modflow.ModflowMnw1,
+            "mnw2": flopy.modflow.ModflowMnw2,
[…]
