### 2022-01-12 288dfa0
Add(examples, autotests, CI, environment): added remainder of pyGSFLOW v1.0.0 data to repository
