+++ autotest/__init__.py
@@ -1,0 +1,1 @@
+# empty init file for coverages

[…]
