+++ RELEASE.md
@@ -1,30 +1,35 @@
 Introduction
 -----------------------------------------------
-pyGSFLOW is a  package to create, read, write, edit and visualize data for 
-GSFLOW models
 
-Contributing
------------------------------------------------
-Bug reports, code contributions, or improvements to the documentation are 
-welcome from the communit. Prior to contributing, please read up on our
-guidelines for [contributing](https://github.com/pygsflow/pygsflow/blob/develop/CONTRIBUTING.md).
+pyGSFLOW is a  package to create, read, write, edit and visualize data for GSFLOW models
+
 
 Documentation
 ------------------------------------------------
+
 pyGSFLOW code documentation is available at [https://pygsflow.github.io/pygsflowdocs/](https://pygsflow.github.io/pygsflowdocs/)
+
+
+Examples
+------------------------------------------------
+
+Example notebooks can be found in the github repository at [https://github.com/pygsflow/pygsflow/tree/master/examples](https://github.com/pygsflow/pygsflow/tree/master/examples)
+
 
 Installation
 ------------------------------------------------
-To install pyGSFLOW version 1.0.0:
-```
-pip install pygsflow
-```
-To update to pyGSFLOW version 1.0.0:
-```
-pip install pygsflow --upgrade
-```
+
+To install pyGSFLOW version 1.0.1:
+
+    pip install pygsflow
+
+To update to pyGSFLOW version 1.0.1:
+
+    pip install pygsflow --upgrade
+
 
 How to Cite
 ------------------------------------------------
+
 [Larsen, J. D., Alzraiee, A., Niswonger, R., 2021, pyGSFLOW v1.0.0: U.S. Geological
 Survey Software Release, 2 July 2021, https://doi.org/10.5066/P9NPZ5AD](https://doi.org/10.5066/P9NPZ5AD)
