+++ .github/workflows/ci.yml
@@ -59,7 +59,7 @@ jobs:
     - name: Install flopy and pygsflow
       shell: bash -l {0}
       run: |
-        pip install https://github.com/modflowpy/flopy/zipball/develop
+        pip install flopy==3.3.4
         pip install .
 
     - name: Setup symbolic link to gfortran on Linux
@@ -84,7 +84,7 @@ jobs:
         cd ..
         pip install pytest
         pip install pytest-cov
-        pip install https://github.com/modflowpy/flopy/zipball/develop
+        pip install flopy==3.3.4
         pip install .
         pytest ./autotest/ --cov=gsflow --cov-report=xml
 
