+++ JOSS/paper.bib
@@ -85,6 +85,45 @@
       doi = {10.3133/tm6A16}
 }
 
+@Article{Hunter:2007,
+  Author    = {{Hunter}, J.~D.},
+  Title     = {Matplotlib: A 2D graphics environment},
+  Journal   = {Computing in Science \& Engineering},
+  Volume    = {9},
+  Number    = {3},
+  Pages     = {90--95},
+  abstract  = {Matplotlib is a 2D graphics package used for Python for
+  application development, interactive scripting, and publication-quality
+  image generation across user interfaces and operating systems.},
+  publisher = {IEEE COMPUTER SOC},
+  doi       = {10.1109/MCSE.2007.55},
+  year      = {2007}
+}
+
+@misc{Kluyver:2016,
+	Author = {{Kluyver}, T., 
+		      and {Ragan-Kelley}, B. 
+			  and {P{\'e}rez}, F.
+			  and {Granger}, B.  
+			  and {Bussonnier}, M.
+			  and {Frederic}, J. 
+			  and {Kelley}, K. 
+			  and {Hamrick}, J.,
+			  and {Grout}, J. 
+			  and {Corlay}, S. 
+			  and {Ivanov}, P. 
+			  and {Avila}, D. 
+			  and {Abdalla}, S. 
+			  and {Willing}, C.},
+	Booktitle = {Positioning and Power in Academic Publishing: Players, Agents and Agendas},
+	Editor = {F. Loizides and B. Schmidt},
+	Organization = {IOS Press},
+	Pages = {87 - 90},
+	Title = {Jupyter Notebooks -- a publishing format for reproducible computational workflows},
+	doi = {10.3233/978-1-61499-649-1-87},
+	Year = {2016}
+}
+
 @misc{Labadie:2006,
       author = {{Labadie}, J.
                 and {Larson}, R.},
@@ -174,6 +213,17 @@
       Title = {MODFLOW-USG version 1: An unstructured grid version of MODFLOW for simulating groundwater flow and tightly coupled processes using a control volume finite-difference formulation},
       Year = 2013,
       doi = {10.3133/tm6A45}
+}
+
+@book{Schroeder:2006,
+      title={The Visualization Toolkit: An Object-oriented Approach to 3D Graphics},
+      author={{Schroeder}, W. 
+	          and {Martin}, K. 
+			  and {Lorensen}, B.
+			  ~ISBN 978-1-930934-19-1.},
+      url={https://books.google.com/books?id=rx4vPwAACAAJ},
+      year={2006},
+      publisher={Kitware}
 }
 
 @article{Volk:2019,
