+++ JOSS/paper.md
@@ -20,136 +20,132 @@ authors:
     ocrid: 0000-0001-6397-2403
     affiliation: 2  
 affiliations:
-  - name: U.S. Geological Survey, California Water Science Center, Sacramento, CA
+  - name: California Water Science Center, United States Geological Survey, Sacramento, CA
     index: 1
-  - name: U.S. Geological Survey, Water Mission Area, Menlo Park, CA
+  - name: Integrated Modeling and Prediction Division, Water Mission Area, United States Geological Survey, Menlo Park, CA
     index: 2  
 date: 19 March 2021
 bibliography: paper.bib
 ---
 
 # Overview
-pyGSFLOW is a python package designed to create new GSFLOW integrated 
-hydrologic models, read existing models, edit model input data, run GSFLOW 
-models, process output, and visualize model data.
+pyGSFLOW is a python package designed to create new GSFLOW integrated hydrologic 
+models, read existing models, edit model input data, run GSFLOW models, process  
+output, and visualize model data.
 
 # Introduction
 Hydrologic modeling has been steadily increasing in complexity over the years. 
 The addition of new types of data sets and the need to represent the interaction 
 between surface water and groundwater systems has been driving this complexity. 
-Ignoring landscape changes in space and time and applying simple boundary 
-conditions within hydrologic models is no longer adequate to address watershed 
-and basin scale issues [@Fatichi:2016]. Instead, integrated hydrologic 
-models (IHMs), that couple governing equations for surface water and groundwater 
-flow, are used to represent feedback mechanisms between these systems.
+Ignoring landscape changes in space and time and applying simple boundary conditions 
+within hydrologic models is no longer adequate to address watershed and basin scale 
+issues [@Fatichi:2016]. Instead, integrated hydrologic models (IHMs), that couple 
+governing equations for surface water and groundwater flow, are used to represent 
+feedback mechanisms between these systems.
 
 Among these IHMs is GSFLOW simulation code [@Markstrom:2008] that 
 simulates surface and subsurface hydrologic processes by integrating the 
 Precipitation Runoff Modeling System (PRMS) [@Markstrom:2015] and 
 MODFLOW [@Harbaugh:2005; @Niswonger:2011] into a single code that 
-simulates feedbacks between the two systems. Because modelers are moving toward 
-simulating greater portions of the hydrologic cycle, larger datasets, from 
-multiple sources are used to parameterize these models. Beyond the scope of 
-example problems, most applied problems require custom workflows and code to 
-process large datasets related to model inputs and outputs. Scripting languages 
-like Python, R, and MATLAB make it easier to process large data sets and 
-provide standard methods that can be used for developing, editing, and properly 
-formatting model input files and for analyzing model output data. These 
-developments have led to major advancements in model reproducibility and 
-improvements in model applicability [@Bakker:2016; @Gardner:2018; @Ng:2018].
+simulates feedbacks between the two processes. Because modelers are moving 
+toward simulating greater portions of the hydrologic cycle, larger datasets, 
+from multiple sources are used to parameterize these models. Beyond the scope 
+of example problems, most applied problems require custom workflows and code 
+to process large datasets related to model inputs and outputs. Scripting languages 
+like Python, R, and MATLAB make it easier to process large data sets and provide 
+standard methods that can be used for developing, editing, and properly formatting 
+model input files and for analyzing model output data. These developments have led 
+to major advancements in model reproducibility and improvements in model 
+applicability [@Bakker:2016; @Gardner:2018; @Ng:2018].
 
 # Statement of need
 GSFLOW model development previously has been a piecemeal approach. Arcpy-GSFLOW 
[…]
