+++ gsflow/modflow/mfag.py
@@ -27,6 +27,8 @@ class ModflowAg(flopy.modflow.ModflowAg):
         dictionary of the irrdiversion block
     irrwell : dict {per: np.recarray}
         dictionary of the irrwell block
+    irrpond : dict {per: np.recarray}
+        dictionary of the irrpond block
     supwell : dict {per: np.recarray}
         dictionary of the supwell block
     extension : str, optional
@@ -44,8 +46,8 @@ class ModflowAg(flopy.modflow.ModflowAg):
     load a ModflowAg file
 
     >>> import gsflow
-    >>> ml = gsflow.modflow.Modflow('awutest')
-    >>> ag = gsflow.modflow.ModflowAg.load('test.awu', ml, nper=2)
+    >>> ml = gsflow.modflow.Modflow('agtest')
+    >>> ag = gsflow.modflow.ModflowAg.load('test.g', ml, nper=2)
 
     """
 
@@ -167,7 +169,16 @@ class ModflowAg(flopy.modflow.ModflowAg):
                     OptionBlock.nested: True,
                     OptionBlock.n_nested: 1,
                     OptionBlock.vars: OrderedDict(
-                        [("accel", OptionBlock.simple_float)]
+                        [
+                            (
+                                "accel",
+                                {
+                                    OptionBlock.dtype: float,
+                                    OptionBlock.nested: False,
+                                    OptionBlock.optional: True,
+                                },
+                            )
+                        ]
                     ),
                 },
             ),
@@ -312,9 +323,34 @@ class ModflowAg(flopy.modflow.ModflowAg):
             nper=nper,
         )
 
+    @property
+    def segment_list(self):
+        """
+        Method to get a unique list of segments from irrdiversion and irrpond
+
+        """
+        segments = []
+        if self.irrdiversion is not None:
+            for _, recarray in self.irrdiversion.items():
+                if np.isscalar(recarray):
+                    continue
+                t = np.unique(recarray["segid"])
+                for seg in t:
+                    segments.append(seg)
+
+            segments = list(set(segments))
+
+        # if pond list exists pop off segments that are routed to ponds
+        if self.pond_list is not None:
+            for seg in self.pond_list["segid"]:
+                if seg in segments:
+                    segments.pop(segments.index(seg))
+
+        return segments
+
     def write_file(self, check=False):
         """
-        Write method for ModflowAwu
+        Write method for ModflowAg
 
         Parameters
         ----------
@@ -329,6 +365,8 @@ class ModflowAg(flopy.modflow.ModflowAg):
             name = self.file_name[0]
             with open(os.path.join(ws, name), "w") as foo:
                 foo.write(self.heading)
+                if not self.heading.endswith("\n"):
+                    foo.write("\n")
 
                 # update options
                 self.options.update_from_package(self)
@@ -994,7 +1032,9 @@ class ModflowAg(flopy.modflow.ModflowAg):
                     nrec = len(t)
 
                     # check if this is block 16a
-                    if isinstance(options.tabfileswell, np.recarray):
+                    if isinstance(
+                        options.tabfileswell, np.recarray
+                    ) or isinstance(options.tabfiles, np.recarray):
                         tf = True
                         well = ModflowAg.get_empty(nrec, block="tabfile_well")
                     else:
@@ -1164,6 +1204,80 @@ class ModflowAg(flopy.modflow.ModflowAg):
             supwell=sup_well,
             nper=nper,
         )
+
+    def plot(self, **kwargs):
+        """
+        Plot 2-D, 3-D, transient 2-D, and stress period list (MfList)
+        package input data
+
+        Parameters
+        ----------
+        **kwargs : dict
+            filename_base : str
+                Base file name that will be used to automatically generate file
[…]

[…]
