### 2022-01-12 3ff0437
update version numbers for consistency
