### 2022-01-12 c4a1080
update(warnings): standardize warning call signature
