+++ JOSS/paper.bib
@@ -132,7 +132,17 @@
       publisher = {Department of Civil Engineering, Colorado State University, Ft. Collins, CO},
       url = {http://modsim.engr.colostate.edu/modsim.php}
 }
-      
+
+@misc{Larsen:2021,
+      author = {{Larsen}, J.~D.
+                and {Alzraiee}, A.
+                and {Niswonger}, R.~G.},
+      title = {pyGSFLOW v1.0.0: U.S. Geological Survey Software Release, 2 July 2021},
+      year = {2021},
+      publisher = {U.S. Geological Survey Software Release},
+      doi = {10.5066/P9NPZ5AD}
+}
+                
 @book{Langevin:2017,
       Author = {{Langevin}, C.~D.
                 and {Hughes}, J.~D.
@@ -226,6 +236,13 @@
       publisher={Kitware}
 }
 
+@misc{U.S. Geological Survey:2022,
+      author = {{U.S. Geological Survey}}
+      title = {USGS water data for the Nation: U.S. Geological Survey National Water Information System database},
+      year = {2022},
+      doi = {10.5066/F7P55KJN}
+}
+           
 @article{Volk:2019,
          author = {{Volk}, J.~M.
                    and {Turner}, M.~A.},
