+++ JOSS/paper.bib
@@ -236,8 +236,8 @@
       publisher={Kitware}
 }
 
-@misc{U.S. Geological Survey:2022,
-      author = {{U.S. Geological Survey}}
+@misc{U.S.~Geological~Survey:2022,
+      author = {{U.S.~Geological~Survey}}
       title = {USGS water data for the Nation: U.S. Geological Survey National Water Information System database},
       year = {2022},
       doi = {10.5066/F7P55KJN}
