### 2022-02-18 9ca479a
updated JOSS paper
