+++ JOSS/paper.bib
@@ -237,7 +237,7 @@
 }
 
 @misc{USGS:2022,
-      author = {{U.S. Geological Survey}}
+      author = {{U.S. Geological Survey}},
       title = {USGS water data for the Nation: U.S. Geological Survey National Water Information System database},
       year = {2022},
       doi = {10.5066/F7P55KJN}
