### 2022-02-18 721c033
formatting updates to paper.md
