+++ JOSS/paper.bib
@@ -137,7 +137,7 @@
       author = {{Larsen}, J.~D.
                 and {Alzraiee}, A.
                 and {Niswonger}, R.~G.},
-      title = {pyGSFLOW v1.0.0: U.S. Geological Survey Software Release, 2 July 2021},
+      title = {pyGSFLOW v1.0.0},
       year = {2021},
       publisher = {U.S. Geological Survey Software Release},
       doi = {10.5066/P9NPZ5AD}
