### 2022-02-18 9ececad
Proofreading updates to paper.bib
