+++ JOSS/paper.bib
@@ -80,7 +80,7 @@
       author = {{Harbaugh}, A.~W.},
       Booktitle = {U.S. Geological Survey Techniques and Methods, book 6},
       Publisher = {U.S. Geological Survey},
-      Title = {MODFLOW-2005, the U.S. Geological Survey modular ground-water model -- the Ground-Water Flow Process},
+      Title = {MODFLOW-2005, the {U.S. Geological Survey} modular ground-water model -- the Ground-Water Flow Process},
       year = 2005,
       doi = {10.3133/tm6A16}
 }
@@ -193,7 +193,7 @@
                    and {Cronkite-Ratcliff}, C.
                    and {Niswonger}, R.~G.
                    and {McKenzie}, J.~M.},
-         title = {GSFLOW–GRASS v1. 0.0: GIS-enabled hydrologic modeling of coupled groundwater–surface-water systems},
+         title = {GSFLOW–GRASS v1.0.0: GIS-enabled hydrologic modeling of coupled groundwater–surface-water systems},
          year = {2018},
          journal = {Geoscientific Model Development},
          volume = 11,
@@ -207,7 +207,7 @@
                 and {Ibaraki}, M.},
       Booktitle = {U.S. Geological Survey Techniques and Methods, book 6},
       Publisher = {U.S Geological Survey},
-      Title = {MODFLOW-NWT, A Newton formulation for MODFLOW-2005},
+      Title = {MODFLOW-NWT, A Newton formulation for {MODFLOW-2005}},
       Year = 2011,
       doi = {10.3133/tm6A37}
 }
@@ -220,7 +220,7 @@
                 and {Hughes}, J.~D.},
       Booktitle = {U.S. Geological Survey Techniques and Methods, book 6},
       Publisher = {U.S Geological Survey},
-      Title = {MODFLOW-USG version 1: An unstructured grid version of MODFLOW for simulating groundwater flow and tightly coupled processes using a control volume finite-difference formulation},
+      Title = {MODFLOW-USG version 1: An unstructured grid version of {MODFLOW} for simulating groundwater flow and tightly coupled processes using a control volume finite-difference formulation},
       Year = 2013,
       doi = {10.3133/tm6A45}
 }
@@ -229,8 +229,7 @@
       title={The Visualization Toolkit: An Object-oriented Approach to 3D Graphics},
       author={{Schroeder}, W. 
 	          and {Martin}, K. 
-			  and {Lorensen}, B.
-			  ~ISBN 978-1-930934-19-1.},
+			  and {Lorensen}, B.~ISBN 978-1-930934-19-1 (HB).},
       url={https://books.google.com/books?id=rx4vPwAACAAJ},
       year={2006},
       publisher={Kitware}
@@ -238,7 +237,7 @@
 
 @misc{USGS:2022,
       author = {{U.S. Geological Survey}},
-      title = {USGS water data for the Nation: U.S. Geological Survey National Water Information System database},
+      title = {USGS water data for the Nation: {U.S. Geological Survey} National Water Information System database},
       year = {2022},
       doi = {10.5066/F7P55KJN}
 }
