+++ JOSS/paper.bib
@@ -229,10 +229,11 @@
       title={The Visualization Toolkit: An Object-oriented Approach to 3D Graphics},
       author={{Schroeder}, W. 
 	          and {Martin}, K. 
-			  and {Lorensen}, B.~ISBN 978-1-930934-19-1 (HB).},
+			  and {Lorensen}, B.},
       url={https://books.google.com/books?id=rx4vPwAACAAJ},
       year={2006},
-      publisher={Kitware}
+      publisher={Kitware},
+      note = {{ISBN:} 978-1-930934-19-1} 
 }
 
 @misc{USGS:2022,
