+++ JOSS/paper.bib
@@ -229,11 +229,10 @@
       title={The Visualization Toolkit: An Object-oriented Approach to 3D Graphics},
       author={{Schroeder}, W. 
 	          and {Martin}, K. 
-			  and {Lorensen}, B.},
+			  and {Lorensen}, B.~ISBN 978-1930934191.},
       url={https://books.google.com/books?id=rx4vPwAACAAJ},
       year={2006},
-      publisher={Kitware},
-      note = {{ISBN:} 978-1-930934-19-1} 
+      publisher={Kitware}
 }
 
 @misc{USGS:2022,
