+++ JOSS/paper.bib
@@ -226,10 +226,10 @@
 }
 
 @book{Schroeder:2006,
-      title={The Visualization Toolkit: An Object-oriented Approach to 3D Graphics},
+      title={The Visualization Toolkit: An Object-oriented Approach to 3D Graphics.~ISBN 978-1-930934-19-1.},
       author={{Schroeder}, W. 
 	          and {Martin}, K. 
-			  and {Lorensen}, B.~ISBN 978-1930934191.},
+			  and {Lorensen}, B.},
       url={https://books.google.com/books?id=rx4vPwAACAAJ},
       year={2006},
       publisher={Kitware}
