### 2022-02-28 #3852 @crvernon
@jlarsen-usgs The following should be addressed in the paper:

- [ ] Line 8.  Capitalize Python.
- [ ] Line 12.  Need citations backing the claims made in the first and second sentences.
- [ ] Line 19.  "Among these IHMs is GSFLOW..." should be "Among these IHMs is the GSFLOW..."
- [ ] Line 23.  "...larger datasets, from" should be "...larger datasets from"
- [ ] Line 32. "GSFLOW model development previously has been a..." should be "GSFLOW model development has previously been..."
- [ ] Line 38.  "integrated hydrologic models" should be "IHMs" - check this throughout the paper
- [ ] Line 51-52.  Use IHMs instead of exploded name
- [ ] Line 131.  should be "United States Geological Survey (USGS)" since it is the first use.

❗ The current license (CC0) is not allowed for use in JOSS as it is not in the list of OSI-approved licenses that we utilize (see https://opensource.org/licenses/alphabetical).  The following is some discussion on the current status of CC0 license use in the open source community if your are interested:  https://opensource.org/faq#cc-zero and https://opensource.org/faq#public-domain.  Please choose a new license to proceed.

Thanks and let me know if you have any questions!
