### 2022-02-28 #3852 @crvernon
@jlarsen-usgs ... BTW, here is the USGS guidance on licensing that may help you:  https://www.usgs.gov/software-management/distribution-usgs-code

This indicates that you should be able to use an alternative option.  Feel free to reach out if you have questions.  Thanks!
