### 2022-02-28 #3852 @jlarsen-usgs
@crvernon

I looked over that guidence earlier today and it looks like "The Unlicense" is the acceptable option that's included in the OSI approved licenses that you sent. I sent out an email earlier today to check with our policy folks to make sure there isn't any guidence that superceedes this. I'll be updating the paper based on your review comments sometime today.

Thank you for help and your review!
