### 2022-03-01 6d1486e
updates to JOSS paper for editorial comments
