+++ JOSS/paper.md
@@ -29,26 +29,29 @@ bibliography: paper.bib
 ---
 
 # Overview
-pyGSFLOW is a python package designed to create new GSFLOW integrated hydrologic 
+pyGSFLOW is a Python package designed to create new GSFLOW integrated hydrologic 
 models, read existing models, edit model input data, run GSFLOW models, process 
 output, and visualize model data.
 
 # Introduction
-Hydrologic modeling has been steadily increasing in complexity over the years. 
-The addition of new types of data sets and the need to represent the interaction 
-between surface-water and groundwater systems has been driving this complexity. 
+Water resources are becoming increasingly scarce due to pressures from competing interests 
+and rapid development [@Garrote:2017; @Lotze:2006; @Mekonnen:2016]. Groundwater models 
+often serve as a tool for exploring the impacts of different water management pressures 
+and conservation scenarios [@Moore:1980]. The traditional groundwater modeling approach 
+has focused on simulating saturated flow in the groundwater system and uses simplified 
+boundary conditions for representing recharge and discharge from the system [@Chiu:2010; Li:2011; Siade:2014]. 
 Ignoring landscape changes in space and time and applying simple boundary conditions 
 within hydrologic models is no longer adequate to address watershed and basin scale 
 issues [@Fatichi:2016]. Instead, integrated hydrologic models (IHMs), that couple 
 governing equations for surface-water and groundwater flow, are used to represent 
 feedback mechanisms between these systems.
 
-Among these IHMs is GSFLOW simulation code [@Markstrom:2008] that simulates surface 
+Among these IHMs is the GSFLOW simulation code [@Markstrom:2008] that simulates surface 
 and subsurface hydrologic processes by integrating the Precipitation Runoff 
 Modeling System (PRMS)  [@Markstrom:2015] and MODFLOW [@Harbaugh:2005; @Niswonger:2011] 
 into a single code that simulates feedbacks between the two systems. Because 
 modelers are moving toward simulating greater portions of the hydrologic cycle, 
-larger datasets, from multiple sources are used to parameterize these models. 
+larger datasets from multiple sources are used to parameterize these models. 
 Beyond the scope of example problems, most applied problems require custom workflows 
 and code to process large datasets related to model inputs and outputs. Scripting 
 languages like Python make it easier to process large data sets and provide standard 
@@ -58,14 +61,14 @@ major advancements in model reproducibility and improvements in model
 applicability [@Bakker:2016; @Gardner:2018; @Ng:2018].
 
 # Statement of need
-GSFLOW model development previously has been a piecemeal approach. Arcpy-GSFLOW 
+GSFLOW model development has previously been a piecemeal approach. Arcpy-GSFLOW 
 scripts [@Gardner:2018] or GSFLOW-GRASS packages have been used to 
 process surface-water input data into model files. PRMS-Python 
 [@Volk:2019] could be used to edit most of the PRMS inputs to 
 GSFLOW. Finally, FloPy [@Bakker:2016; @Bakker:2021] could be used to edit most 
 of the MODFLOW inputs to GSFLOW. This approach is not tightly coupled and still 
 requires manual edits and additional external scripts to edit, run models, and 
-process output data. Because of the complexity of integrated hydrologic models 
+process output data. Because of the complexity of IHMs 
 and the need for model reproducibility, a single integrated scripting package 
 will help standardize and streamline model development and calibration. 
 
@@ -80,7 +83,7 @@ operations networks, and MODFLOW [@Harbaugh:2005] model packages and binds them
 into a single integrated model instance. pyGSFLOW leverages features from FloPy, 
 an existing Python package for the MODFLOW suite of groundwater modeling software
[…]
