+++ JOSS/paper.bib
@@ -34,6 +34,18 @@
       doi = {10.5066/F7BK19FH}
 }
 
+@article{Chiu:2010,
+         author = {{Chiu}, Y.,
+		           and {Nishikawa}, T.
+				   and {Yeh}, W.~-G.},
+		 title = {Optimal pump and recharge management model for nitrite removal in the Warren Groundwater Basin, California},
+		 year = {2010},
+		 journal = {Journal of Water Resources Planning and Management},
+		 volume = {136},
+		 number = {3},
+		 doi = {10.1061/(ASCE)WR.1943-5452.0000034}
+}
+
 @article{Fatichi:2016,
          author = {{Fatichi}, S.
                    and {Vivoni}, E.~R.
@@ -74,6 +86,16 @@
          volume = 109,
          pages = {41 - 53},
          doi = {10.1016/j.envsoft.2018.07.020}
+}
+
+@article{Garote:2017,
+         author = {{Garote}, L.},
+		 title = {Managing water resources to adapt to climate change: facing uncertainty and scarcity in a changing context},
+		 journal = {Water Resources Management},
+		 year = {2017},
+		 volume = {31},
+		 pages = {2951 - 2963},
+		 doi = {10.1007/s11269-017-1714-6}
 }
 
 @book{Harbaugh:2005,
@@ -157,6 +179,30 @@
       doi = {10.3133/tm6A55}
 }
 
+@misc{Li: 2011,
+	  author = {{Li}, Z.
+	            and {Martin}, P.},
+	  title = {Geohydrology, simulation of regional groundwater flow, and assessment of water management strategies, Twentynine Palms area, California},
+	  publisher = {U.S. Geological Survey Scientific Investigations Report},
+      year = {2011},
+	  url = {https://pubs.usgs.gov/sir/2010/5249/}
+}
+
+@article{Lotze:2006,
+         author = {{Lotze-Campen}, H.
+		           and {Muller}, C.
+				   and {Bondeau}, A.
+				   and {Stefanie}, R.
+				   and {Popp}, A.
+				   and {Lucht}, W.},
+		 title = {Global food demand, productivity growth, and the scarcity of land and water resources: a spatially explicit mathematical programming approach},
+		 journal = {Agricultural Economics},
+         year = {2006},
+         volume = 39,
+         pages = {325 - 338},
+         doi = {10.1111/j.1574-0862.2008.00336.x}
+}			  
+
 @book{Markstrom:2008,
       Author = {{Markstrom}, S.~L.
                 and {Niswonger}, R.~G.
@@ -183,6 +229,25 @@
       Title = {PRMS-IV, the precipitation-runoff modeling system, version 4},
       Year = 2015,
       doi = {10.3133/tm6B7}
+}
+
+@article{Mesfin:2016,
+         author = {{Mesfin}, M.~M.
+		           and {Hoekstra}, A.~Y.},
+		 title = {Four billion people facing severe water scarcity},
+		 year = {2016},
+		 journal = {Science Advances},
+		 volume = {2},
+		 number = {2},
+		 doi = {https://doi.org/10.1126/sciadv.1500323}
+}
+
+@misc{Moore:1980,
+      author = {{Moore}, J.~E.},
+	  title = {Ground-water models for water resources planning},
+	  year = {1980},
+	  publisher = {U.S. Geological Survey Open-file report 80-581},
+	  doi = {10.3133/ofr80581}
 }
 
 @article{Ng:2018,
@@ -241,7 +306,19 @@
       year = {2022},
       doi = {10.5066/F7P55KJN}
 }
-           
+
+@misc{Siade:2014,
+      author = {{Siade}, A.~J.
+	            and {Nishikawa}, T.
[…]
