+++ JOSS/paper.bib
@@ -179,7 +179,7 @@
       doi = {10.3133/tm6A55}
 }
 
-@misc{Li: 2011,
+@misc{Li:2011,
 	  author = {{Li}, Z.
 	            and {Martin}, P.},
 	  title = {Geohydrology, simulation of regional groundwater flow, and assessment of water management strategies, Twentynine Palms area, California},
