+++ JOSS/paper.bib
@@ -88,8 +88,8 @@
          doi = {10.1016/j.envsoft.2018.07.020}
 }
 
-@article{Garote:2017,
-         author = {{Garote}, L.},
+@article{Garrote:2017,
+         author = {{Garrote}, L.},
 		 title = {Managing water resources to adapt to climate change: facing uncertainty and scarcity in a changing context},
 		 journal = {Water Resources Management},
 		 year = {2017},
@@ -231,15 +231,15 @@
       doi = {10.3133/tm6B7}
 }
 
-@article{Mesfin:2016,
-         author = {{Mesfin}, M.~M.
+@article{Mekonnen:2016,
+         author = {{Mekonnen}, M.~M.
 		           and {Hoekstra}, A.~Y.},
 		 title = {Four billion people facing severe water scarcity},
 		 year = {2016},
 		 journal = {Science Advances},
 		 volume = {2},
 		 number = {2},
-		 doi = {https://doi.org/10.1126/sciadv.1500323}
+		 doi = {10.1126/sciadv.1500323}
 }
 
 @misc{Moore:1980,
