### 2022-03-01 636aa63
update paper.md and paper.bib in response to editorial reviews
