+++ JOSS/paper.bib
@@ -46,6 +46,15 @@
 		 doi = {10.1061/(ASCE)WR.1943-5452.0000034}
 }
 
+@book{Davie:2019,
+      authors = {{Davie}, T.
+	             and {Quinn}, N.~W.},
+	  Booktitle = {Fundamentals of Hydrology},
+	  year = {2019},
+	  publisher = {Taylor & Francis},
+	  doi = {10.4324/9780203798942}
+}	  
+
 @article{Fatichi:2016,
          author = {{Fatichi}, S.
                    and {Vivoni}, E.~R.
@@ -240,14 +249,6 @@
 		 volume = {2},
 		 number = {2},
 		 doi = {10.1126/sciadv.1500323}
-}
-
-@misc{Moore:1980,
-      author = {{Moore}, J.~E.},
-	  title = {Ground-water models for water resources planning},
-	  year = {1980},
-	  publisher = {U.S. Geological Survey Open-file report 80-581},
-	  doi = {10.3133/ofr80581}
 }
 
 @article{Ng:2018,
