### 2022-03-01 ccd01a3
fix typo in paper.md
