### 2022-03-01 c086637
update spacing between text and ref
