+++ JOSS/paper.bib
@@ -47,11 +47,12 @@
 }
 
 @book{Davie:2019,
-      authors = {{Davie}, T.
-	             and {Quinn}, N.~W.},
-	  Booktitle = {Fundamentals of Hydrology},
+      author = {{Davie}, T.
+	            and {Quinn}, N.~W.},
+	  Booktitle = {Fundamentals of Hydrology (3rd ed.)},
+	  title = {Fundamentals of Hydrology},
 	  year = {2019},
-	  publisher = {Taylor & Francis},
+	  publisher = {Routledge},
 	  doi = {10.4324/9780203798942}
 }	  
 
