### 2022-03-01 8499f32
update(paper.bib, paper.md): merge in JOSS review changes
