### 2022-03-01 #3852 @jlarsen-usgs
@crvernon

I've updated both the license and the paper based on your review comments. I've also updated the first paragraph of the introduction for better clarity based on your review comment about "Line 12".
