### 2022-03-02 c1a8936
fix minor grammatical typo in paper.md #openjournals/joss-reviews#3825
