### 2022-03-07 #3852 @crvernon
👏 Nice work on the revisions @jlarsen-usgs

Please check on the above missing and invalid DOIs ⬆️ and see if we can get those fixed up.  Then I will be ready to move to the next step in this process.  Also, you may want to check the spelling of "hydraulicly" on Line 21; may need two l's.
