### 2022-03-07 6d89c9c
test update for dropped parenthesis in DOI
