+++ JOSS/paper.bib
@@ -43,7 +43,7 @@
 		 journal = {Journal of Water Resources Planning and Management},
 		 volume = {136},
 		 number = {3},
-		 doi = {10.1061/(ASCE)WR.1943-5452.0000034}
+		 doi = {10.1061/{(ASCE)WR.1943-5452.0000034}}
 }
 
 @book{Davie:2019,
