### 2022-03-07 #3852 @jlarsen-usgs
@crvernon

I updated "hydraulicly to hydraulically" in the paper. Thanks for catching this!

I checked the DOI. In the PDF that's generated the DOI is listed as `10.1061/(ASCE)WR.1943-5452.0000034`, which is the correct (valid) DOI. The `check references` method seems either to not be recognising or ignoring the parenthesis around (ASCE) which is causing it to report falsely that the DOI is invalid. See reference "Chiu, Y., Nishikawa, T., Yeh, W.G. ..." in the PDF of the paper for confirmation.
