+++ JOSS/paper.bib
@@ -195,7 +195,7 @@
 	  title = {Geohydrology, simulation of regional groundwater flow, and assessment of water management strategies, Twentynine Palms area, California},
 	  publisher = {U.S. Geological Survey Scientific Investigations Report},
       year = {2011},
-	  url = {https://pubs.usgs.gov/sir/2010/5249/}
+	  url = {10.3133/sir20105249}
 }
 
 @article{Lotze:2006,
