+++ JOSS/paper.bib
@@ -195,7 +195,7 @@
 	  title = {Geohydrology, simulation of regional groundwater flow, and assessment of water management strategies, Twentynine Palms area, California},
 	  publisher = {U.S. Geological Survey Scientific Investigations Report},
       year = {2011},
-	  url = {10.3133/sir20105249}
+	  doi = {10.3133/sir20105249}
 }
 
 @article{Lotze:2006,
