### 2022-03-23 #3852 @jlarsen-usgs
@crvernon

Reference check passes now. All DOIs are present and recognized.
