+++ examples/data/geospatial/resampled_90m/dem_90m_median.txt
@@ -1,0 +1,77 @@
+2.259313720703125000e+03  2.245037231445312500e+03  2.227583862304687500e+03  2.212610717773437500e+03  2.205684448242187500e+03  2.203361816406250000e+03  2.202472534179687500e+03  2.202354248046875000e+03  2.201894409179687500e+03  2.201364379882812500e+03  2.198863037109375000e+03  2.191127685546875000e+03  2.176597167968750000e+03  2.155254760742187500e+03  2.145744628906250000e+03  2.142468750000000000e+03  2.139946289062500000e+03  2.146671020507812500e+03  2.147722656250000000e+03  2.145794067382812500e+03  2.140697631835937500e+03  2.140646484375000000e+03  2.142228637695312500e+03  2.143791015625000000e+03  2.143520141601562500e+03  2.141890991210937500e+03  2.133023315429687500e+03  2.115053588867187500e+03  2.108263549804687500e+03  2.107734497070312500e+03  2.107791137695312500e+03  2.107568603515625000e+03  2.101805664062500000e+03  2.103986816406250000e+03  2.089063598632812500e+03  2.067797485351562500e+03  2.080410156250000000e+03  2.109440795898437500e+03  2.125417114257812500e+03  2.127907714843750000e+03  2.129312500000000000e+03  2.126425903320312500e+03  2.122780883789062500e+03  2.119391235351562500e+03  2.118141357421875000e+03  2.118974609375000000e+03  2.121652221679687500e+03  2.127846557617187500e+03  2.135104614257812500e+03  2.142392578125000000e+03  2.144771240234375000e+03  2.142413574218750000e+03  2.137193969726562500e+03  2.135944824218750000e+03  2.130887084960937500e+03  2.124101684570312500e+03  2.126769409179687500e+03  2.126716430664062500e+03  2.111525390625000000e+03  2.097460327148437500e+03  2.089317016601562500e+03  2.091913208007812500e+03  2.099032714843750000e+03  2.104851196289062500e+03  2.108700195312500000e+03  2.107671142578125000e+03  2.100731079101562500e+03  2.088385620117187500e+03  2.076569335937500000e+03  2.070145874023437500e+03  2.062324829101562500e+03  2.055438964843750000e+03  2.048507812500000000e+03  2.041641540527343750e+03  2.037839111328125000e+03  2.039085083007812500e+03  2.036995483398437500e+03  2.035979675292968750e+03  2.040194580078125000e+03  2.038947814941406250e+03  2.021797485351562500e+03  2.000340515136718750e+03  1.990402587890625000e+03
[…]
