+++ examples/data/sagehen/50m_tutorials/sagehen_50m_flowdir.txt
@@ -1,0 +1,138 @@
+1  1  1  1  1  1  1  1  1  1  1  1  16  16  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  16  16  16  16  1  1  1  1  1  16  16  16  16  1  1  1  1  1  1  1  1  1  1  16  16  16  1  1  1  1  16  1  1  1  1  16  16  16  16  16  16  16  16  16  1  1  1  1  1  1  1  16  16  16  16  16  8  8  16  16  16  16  1  1  1  1  1  1  1  1  1  16  16  16  1  1  1  1  1  1  16  16  16  16  8  8  16  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  16  16  1  1  1  16  16  16  2  1  1  1  1  1  4
+128  128  128  128  128  128  64  64  128  64  64  64  64  64  64  64  64  1  1  128  1  1  1  1  1  1  128  128  64  64  16  32  32  64  128  128  64  64  64  32  64  32  64  64  64  64  1  1  128  128  64  64  64  64  64  64  64  64  64  64  2  1  1  64  64  16  16  32  32  64  32  64  32  128  1  128  128  128  128  64  64  32  32  16  16  16  16  16  16  16  32  128  128  128  64  64  64  128  128  64  64  64  32  128  128  128  128  128  64  64  32  32  16  16  16  16  16  128  128  128  128  128  128  128  1  1  128  128  128  128  128  128  128  64  64  64  128  128  64  64  32  32  1  1  1  1  128  1  64
+128  128  128  128  128  128  128  128  64  64  64  64  64  64  64  64  128  128  1  1  1  1  1  1  1  128  128  128  128  64  32  16  32  128  1  128  128  64  64  32  32  64  64  64  64  128  1  1  128  128  64  64  64  64  64  64  64  128  64  64  1  1  128  64  32  32  32  32  32  64  64  32  128  128  128  128  128  128  128  64  64  64  32  16  16  16  16  16  16  16  64  64  128  64  64  64  128  128  64  64  64  64  128  1  1  1  128  128  128  64  64  32  16  16  16  16  16  64  128  128  128  128  1  1  1  128  128  128  128  128  128  128  128  64  64  64  128  64  64  64  32  1  1  1  1  128  128  128  64
+128  128  128  128  128  128  128  128  128  64  64  64  64  64  64  64  128  128  1  1  1  1  1  1  128  128  128  128  64  64  32  16  32  1  1  1  128  64  64  64  64  32  32  64  64  1  1  128  128  64  64  64  64  64  64  64  64  128  64  2  1  128  128  64  32  16  32  32  32  64  64  1  128  128  128  128  128  128  128  64  128  1  64  32  16  16  16  16  16  16  16  64  64  64  64  64  64  64  64  64  128  1  2  1  1  1  128  1  128  128  64  64  32  16  16  16  16  16  64  128  1  1  1  1  1  128  128  1  1  128  128  128  128  128  128  64  64  64  64  64  64  1  1  1  128  128  1  1  64
[…]
