+++ gsflow/builder/flow_accumulation.py
@@ -1,0 +1,1750 @@
+import numpy as np
+from collections import OrderedDict
+import random
+from ..utils.sfr_renumber import Topology
+from ..utils import gsflow_io
+from . import Defaults
+from flopy.modflow import ModflowSfr2
+import shapefile
+import inspect
+
+
+class FlowAccumulation(object):
+    """
+    Class to perform flow accumulation on a DEM (raster) or resampled to
+    a model grid
+
+    Parameters:
+    ----------
+    data : np.ndarray
+        two dimensional numpy array of dem data
+    xcenters : np.ndarray
+        a two dimensional array of x coordinate cell centers
+    ycenters : np.ndarray
+        two dimensional array of y coordinate cell centers
+    acc_type : str
+        flow accumulation type, currently supported options are "d8"
+    hru_type : np.array
+        optional numpy array of hru type numbers that can be used as
+        a masking array to exclude lakes, swales, etc...
+        0 == inactive
+        1 == land (included in flow accumulation)
+        2 == lake (excluded from flow accumulation)
+        3 == swale (excluded from flow accumulation)
+    closed_basin : bool
+        method to indicate that basin is closed without considering lake
+        flow. If true hru_type 2 is used in the flow direction calculations.
+        False ignores hru_type 2.
+    flow_dir_array : np.ndarray
+        previously calculated flow direction array of dimension nrow, ncol
+        that can be supplied to FlowAccumulation
+    verbose : bool
+        flag to print verbose output
+
+    """
+
+    def __init__(
+        self,
+        data,
+        xcenters,
+        ycenters,
+        acc_type="d8",
+        hru_type=None,
+        closed_basin=False,
+        flow_dir_array=None,
+        verbose=False,
+    ):
+
+        self._defaults = Defaults().to_dict()
+
+        # set the fa type
+        self._acc_type = acc_type
+        # flow directions vectors set up as top to bot, left to right
+        self._d8_vectors = np.array([32, 64, 128, 16, 1, 8, 4, 2], dtype=int)
+        self._quiver_u = {
+            32: -1,
+            64: 0,
+            128: 1,
+            16: -1,
+            1: 1,
+            8: -1,
+            4: 0,
+            2: 1,
+            -1: np.nan,
+            -2: np.nan,
+        }
+        self._quiver_v = {
+            32: 1,
+            64: 1,
+            128: 1,
+            16: 0,
+            1: 0,
+            8: -1,
+            4: -1,
+            2: -1,
+            -1: np.nan,
+            -2: np.nan,
+        }
+        self._inner_idx = []
+        self._graph = OrderedDict()
+        self._visited = []
+        self._stack = []
+        self._dest = None
+        self._solved_flats = {}
+
+        # adds an extra row and col of data
+        data, xcenters, ycenters, hru_type = self._buffer_data(
+            data, xcenters, ycenters, hru_type
+        )
+
+        self._offset = data.shape[1]
+        self._offsets = np.array(
+            [
+                (-1 * self._offset) - 1,
+                (-1 * self._offset),
+                (-1 * self._offset) + 1,
+                -1,
+                1,
+                self._offset - 1,
+                self._offset,
+                self._offset + 1,
+            ]
+        )
+        self._shape = data.shape
+        self._data = data.ravel()
+        self._xcenters = xcenters.ravel()
+        self._ycenters = ycenters.ravel()
+        self._hru_type = hru_type.ravel()
+        self._flow_directions = np.ones((self._data.size,)) * -1
+        self._closed_basin = closed_basin
+
+        if flow_dir_array is None:
+
+            self._flow_directions = np.ones((self._data.size,)) * -1
+            self._closed_basin = closed_basin
[…]
