+++ gsflow/builder/make_json_defaults.py
@@ -111,7 +111,7 @@ defaults = {
         "csv_output_file": {"dtype": 4, "record": "gsflow.csv"},
         "gsflow_output_file": {"dtype": 4, "record": "gsflow.out"},
         "executable_desc": {"dtype": 4, "record": "GSFLOW model"},
-        "executable_model": {"dtype": 4, "record": "prmsIV.exe"}
+        "executable_model": {"dtype": 4, "record": "prmsIV.exe"},
     },
     "parameter": {
         "dimensions": {
