### 2022-03-24 83be62e
update(test_024.py): update to check for rasterio installation
