### 2022-03-25 c2ef593
fixed(plot_discretization): added set method for matplotlib kwargs
* update warnings call in param_base.py
