+++ examples/frontiers/sagehen_90m.py
@@ -1,0 +1,648 @@
+import os
+import utm
+import platform
+import flopy
+import numpy as np
+import shapefile
+import matplotlib.pyplot as plt
+import matplotlib
+import datetime
+import matplotlib.dates as mdates
+from dateutil.relativedelta import relativedelta
+from flopy.utils import Raster
+from flopy.plot import styles, PlotMapView
+from gsflow import GsflowModel, PrmsModel, PrmsData
+from gsflow.builder import (
+    GenerateFishnet,
+    ModflowBuilder,
+    ControlFileBuilder,
+    PrmsBuilder,
+    FlowAccumulation
+)
+import gsflow.builder.builder_utils as bu
+from datetime import date
+from datetime import timedelta
+from dateutil.parser import parse
+
+
+def nash_sutcliffe_efficiency(qsim, qobs, flg):
+    if flg:
+      qsim = np.log(qsim)
+      qobs = np.log(qobs)
+    qsim[np.isinf(qsim)] = np.nan
+    qobs[np.isinf(qobs)] = np.nan
+    numerator = np.nansum((qsim - qobs) ** 2)
+    denominator = np.nansum((qobs - np.nanmean(qobs)) ** 2)
+    nse = 1 - (numerator / denominator)
+    return nse
+
+
+def build_lut(f, dtype=int):
+    d = {}
+    with open(f) as foo:
+        for line in foo:
+            temp = line.strip().split("#")[0]
+            if not temp:
+                continue
+            else:
+                l = temp.split(":")
+                d[dtype(l[0])] = float(l[1])
+    return d
+
+
+if __name__ == "__main__":
+    sample_grid = True
+    # set file names here
+    ws = os.path.abspath(os.path.dirname(__file__))
+    iws = os.path.join(ws, "..", "data", "geospatial")
+    ows = os.path.join(ws, "temp")
+    if not os.path.exists(ows):
+        os.mkdir(ows)
+
+    dem_file = os.path.join(iws, 'dem.img')
+    pour_point_file = os.path.join(iws, "model_points.shp")
+
+    resampled_dem = os.path.join(ows, "sagehen_90m_med.txt")
+
+    stream_threshold = 810000 # m3 of drainage area
+    cellsize = 90
+
+    # generate a "Fishnet"
+    modelgrid = GenerateFishnet(dem_file, xcellsize=cellsize, ycellsize=cellsize)
+
+    # resample DEM to the model grid using minimum elevation
+    if sample_grid:
+        raster = Raster.load(dem_file)
+        dem = raster.resample_to_grid(
+            modelgrid,
+            band=raster.bands[0],
+            method="median",
+            multithread=True,
+            thread_pool=12
+        )
+        np.savetxt(resampled_dem, dem, delimiter="  ")
+    else:
+        dem = np.genfromtxt(resampled_dem)
+
+    hru_type = np.ones((modelgrid.nrow, modelgrid.ncol), dtype=int)
+    fa = FlowAccumulation(dem,
+                          modelgrid.xcellcenters,
+                          modelgrid.ycellcenters,
+                          hru_type=hru_type, verbose=True)
+
+    flow_dir = fa.flow_directions(dijkstra=True)
+    qx, qy = fa.get_vectors
+
+    flow_acc = fa.flow_accumulation()
+
+    # read in pour point from shapefile and set watershed boundary
+    with shapefile.Reader(pour_point_file) as r:
+        shape = r.shape(0)
+        pour_point = shape.points
+        pour_point[0][1] -= 20
+        pour_point[0][0] -= 20
+
[…]
