### 2022-03-31 069567a
update(make_streams): bugfix for when hru_type is user defined.
* prior to fix basin outlet could circularly route creating infinite numeric loop
