### 2022-03-31 260152a
update(flow_accumulation.py): create more robust strtop calculation
