### 2022-04-01 af1c768
fix(get_cascades): update for user supplied hru_type array
* remove potential circular routing at outlet due to hru_type array
