### 2022-04-04 7f48734
update(test_020.py): update test for ControlFileDefaults.add_record() changes
