### 2022-04-05 f361bda
update(PrmsBuilder): update to allow user to build without SFR information
