### 2022-04-06 #3852 @jlarsen-usgs
@crvernon no worries, the publication can wait until you've got the time. Just tag me in a comment when you're ready to start moving this forward.
