### 2022-04-07 a6acede
update(load_from_file): update to allow user to load model without data file
