### 2022-04-14 bbab0da
fix(load_from_file()): remove duplicate parameters from param_list variable for proper indexing and removal
