### 2022-04-18 #3852 @jlarsen-usgs
@crvernon

I've put together the github release and created a Zenodo archive for this publication. The doi is [10.5281/zenodo.6468138](https://doi.org/10.5281/zenodo.6468138)
