### 2022-04-18 #3852 @jlarsen-usgs
@crvernon

I just read through the proof and found a typo that should be corrected before final acceptance and publication.  "gsflow_coef" should be "gwflow_coef". I'll correct this right now.
