### 2022-04-19 f116c8e
updates(): usgs code review updates
* update sagehen_50m.py to avoid file access errors in GSFLOW
* update rasters.py, remove Rasters class and create child of flopy.utils.Raster
* added Numba based resampling mode to Raster.resample_to_grid
* added autotests for Raster class
* linting
