+++ gsflow/utils/rasters.py
@@ -3,167 +3,668 @@ rasters.py includes code to work with raster datasets,
 particularly to sample raster using a PrmsDiscretization
 object and then use it as input data
 """
+import queue
+import threading
 import numpy as np
-import inspect
-from .gsflow_io import _warning
+import flopy.utils
+import warnings
+
+warnings.simplefilter("ignore", RuntimeWarning)
+
+try:
+    from numba import jit, prange
+
+    ENABLE_JIT = True
+except ImportError:
+    from . import numba
+
+    ENABLE_JIT = False
 
 
-class Rasters(object):
+class Raster(flopy.utils.Raster):
     """
-    DEPRECATED: please use flopy.utils.Rasters, will be removed in version
-    1.1.0
-
-    Raster object which allows the users to snap a raster to a grid.
+    Overloaded flopy.utils.Raster class. This class is used to resample
+    rasters and the overload allows for multiprocessing using the "ray"
+    python library
 
     Parameters
     ----------
-        raster : osgeo.gdal.Dataset object
+    array : np.ndarray
+        a three dimensional array of raster values with dimensions
+        defined by (raster band, nrow, ncol)
+    bands : tuple
+        a tuple of raster bands
+    crs : int, string, rasterio.crs.CRS object
+        either a epsg code, a proj4 string, or a CRS object
+    transform : affine.Affine object
+        affine object, which is used to define geometry
+    nodataval : float
+        raster no data value
+    rio_ds : DatasetReader object
+        rasterIO dataset Reader object
+
+    Notes
+    -----
+
+
+    Examples
+    --------
+    >>> from gsflow.utils import Raster
+    >>>
+    >>> rio = Raster.load("myraster.tif")
 
     """
 
-    def __init__(self, raster):
-        msg = "Rasters is Deprecated and will be removed in v.1.1.0"
-        _warning(
-            msg,
-            inspect.getframeinfo(inspect.currentframe()),
-            DeprecationWarning,
-        )
-        self.raster = raster
-        self._band_data = None
-        self._xpoints = None
-        self._ypoints = None
-        self._xypoints = None
-
-    @property
-    def extent(self):
-        """
-        Returns
-        -------
-            (xmin, xmax, ymin, ymax)
-
-        """
-        return (
-            self.raster.bounds.left,
-            self.raster.bounds.right,
-            self.raster.bounds.bottom,
-            self.raster.bounds.top,
+    def __init__(
+        self,
+        array,
+        bands,
+        crs,
+        transform,
+        nodataval,
+        driver="GTiff",
+        rio_ds=None,
+    ):
+        super(Raster, self).__init__(
+            array, bands, crs, transform, nodataval, driver, rio_ds
         )
 
     @property
-    def xpoints(self):
+    def bounds(self):
         """
-        Returns
-        -------
-            Cell centered x points for raster values
-
+        Returns a tuple of xmin, xmax, ymin, ymax boundaries
         """
-        if self._band_data is not None:
-            if self._xpoints is None:
-                xmin, xmax = self.extent[0:2]
-                ynum, xnum = self._band_data.shape
-                t = np.linspace(xmin, xmax, xnum)
-                self._xpoints = np.tile(t, (ynum, 1))
-
-            return self._xpoints
+        return super(Raster, self).bounds
 
     @property
-    def ypoints(self):
+    def bands(self):
         """
-        Returns
-        -------
-            Cell centered y points for raster values
-
+        Returns a tuple of raster bands
         """
-        if self._band_data is not None:
-            if self._ypoints is None:
-                ymin, ymax = self.extent[2:4]
-                ynum, xnum = self._band_data.shape
-                t = np.linspace(ymin, ymax, ynum)
[…]

[…]
