### 2022-04-19 31f620d
update(rasters.py): update call to fake "jit" decorator
