### 2022-04-19 b6a822e
update(jit): update jit decorator for kwargs
