### 2022-04-19 8446859
update jit decorator for optional numba
