+++ gsflow/utils/numba.py
@@ -2,10 +2,13 @@
 Class to trick python into installing without numba package support.
 
 """
+from functools import wraps
 
 
-def jit(func, **kwargs):
-    def wrapper(*args, **kwargs):
-        func(*args, **kwargs)
-
+def jit(**kwargs):
+    def wrapper(func):
+        @wraps(func)
+        def decorated_func(*args, **kwargs):
+            return func(args, **kwargs)
+        return decorated_func
     return wrapper
