### 2022-04-19 9b7becd
python3.6 compatibility update for rasters.py.
Add python 3.10 to autotest matrix
