+++ .github/workflows/ci.yml
@@ -14,7 +14,7 @@ jobs:
       fail-fast: false
       matrix:
         os: [ubuntu-latest, windows-latest]
-        python-version: [3.9, 3.8, 3.7, 3.6]
+        python-version: [3.10, 3.9, 3.8, 3.7, 3.6]
         run-type: [std]
         test-pth: ["./autotest"]
     defaults:
