### 2022-04-20 #3852 @jlarsen-usgs
@arfon

Hello, the doi link for the paper [https://joss.theoj.org/papers/10.21105/joss.03852](https://joss.theoj.org/papers/10.21105/joss.03852) is showing a 404 not found message from github pages. The pdf download is available however.
